Add HD-only drugs section on letters
Where a patient has prescriptions which are marked
as administer on HD only then display these
on a letter in a separate section entitled
'Drugs to give on Haemodialysis'.
When these HD drugs are changed they are bolded in the ususal way.
When they are terminated they appear in the
Stopped Medications list n the letter in the usual way.