Remove foundation-rails gem
Move js and css files from the gem to vendor/ so we can gradually start commenting-out foundation grid/form etc css classes so our new tailwindcss equivalents can get a look-in.
Foundation 5 is fairly old, and the decision has been made not to migrate to version 6, which is not getting much attention (a handful of commit in the last 6 months) and would require a fair about of effort in any case.