Add an optional drug when creating a vaccination event
When recording a vaccination event, allow an optional
drug to be recorded. The list of drugs that can be selected is taken form the drugs table where the drug has a type of 'vaccine'.
The name of the drug is stored as a string, rather than a reference to the drug id in order to fit with the jsonb paradigm.