Refactor HD session date and time storage
In order to support dialysing overnight, and to support incoming sessions via the API, instead of storing session start date, start time and end time, store started_at and stopped_at datetimes.
To achieve this we have to change the HD session form a bit to allow eg start date and time to be entered separately and combined when saved, and to ask the user if the patient dialysed overnight, in which case we add one day to the stopped_at datetime.