Allow a user to create a 'medication review' event
This event signifies that the user is happy the the list of
drugs in the Current list is up-to-date.

From the prescriptions page a user can click Medication Review
and if they accept the alert, a new Medication::Review event is
created in the events table and the latest review date
and reviewer name is displayed at the top of the prescriptions list.

The latest review (date and reviewer) is also displayed on the
letters under the medications, and on drug lists.

It is also possible to see/create Medication Review events from the
Events section for a patient, because a Medication::Review is a
type of event.