Add bone and anemia patient 'perspectives'
Perspectives are a 'view' onto a patient from a certain angle, currently bone and anemia. They are similar to MDMs in that they aggregate patient data, but they do it
in a more focused and less multidisciplinary way.

Currently perspectives pages are mostly hard-coded, but eventually could be data driven.

Remove the pathology charts tab