Introduce HL7 patient location strategies
Different hospitals require different strategies for locating a patient
referenced in an incoming HL7 message. The strategy to use can be
configured in Renalware.config.hl7_patient_locator_strategy.

At MSE the assigning authority next to the hospital number in the HL7 message
determines what db column (e.g. local_patient_id_2) to search in to find a match.
DOB is also included in the search.
This is the :dob_and_any_nhs_or_assigning_auth_number_spec strategy.

At KCH things are simpler as the assigning authority for any hosp number
is 'PAS Number' (and not eg 'KCH') so we can search for that value
across all local_patient_id* columns. DOB is not used.
This is the :simple strategy.