Bypass CSV COPY for ODS gps import
Rather than using COPY (requiring the CSV file to be on the same disk) insert the downloaded CSV data in a new feed_gps table in batches, and change the SQL fn upserts GPs  to read from that table. This allows importing where the database is not on the same server, or in a different docker container etc.