Allow a user to be 'banned'
Allow a superadmin when editing a user in the
the admin -> users list to set the user to be banned
(notes can be added). Once banned a user cannot login (and
if already logged in will be thrown out). A message is displayed to
them when the attempt to login, explaining that they have been
actively blocked from logging in.