Calculate and store 2 and 5 year KFRE pathology results
On receipt of an ACR (or PCR - ecact code is configurable) pathology result,
calculate the 2-year and 5-year KFRE (Kidney Failure Risk Equation)
using a 4-variable equation based on the sex, age, ACR and EGFR.

Use a wisper pathology listener to ko hook into incoming HL7 messaes
to check for an ACR.

If we manage to generate the KFRE values (there is an ACR and the other
3 variables are present) then create a new observation request with the
descriptio code 'KFRE' (create the description row if missing) and
two observatrions for each of the 2 and 5 year results - use the
observation description codes in Renalware.config (e.g. KFRE2), and create these
descriptions if  missing.

References:
The equation is specified here:
https://www.nice.org.uk/guidance/ng203/chapter/recommendations#4-variable-kidney-failure-risk-equation
Online calculator:
https://www.kidney.org/professionals/kdoqi/gfr_calculator to verify.