Allow an admin to change the hospital site a user works at
Allow an admin to change (or set to 'None') the hospital site a user works at. This is partly for with the Barts branch, but may be useful for other sites in terms of presenting relevant dashboard content, abnd for prepopulating filter dropdowns.
Initially the users's site is only editable by an admin (as at Barts), but perhaps we should allow that at other sites.

Unlike at Barts where the user's site determines what patients they see (part of the Heroic confidentiality setup), we do not generally want to do this at other sites as it leads to confusion when a patient or user is assigned to wrong site.