Add as-you-type filtering to the admin drugs list
Proof of concept of a simple addition we can roll out to various filters to avoid the user having to click the Search button when filtering - it will filter as you type/change dropdowns etc , automatically reflecting updating the results.

Here is is applied to the admin drugs list (where you can search by drug name) and provided where are no issues when deployed, the concept can be applied elsewhere.

Uses turboframes and a stimulus controller.