Improve download soft-deletion
Rather than soft-delete with #destroy +acts_as_paranoid, which will cause the related activestorage attachment to purged from disk, here in the controller action update the delete_at manually to do a very explicit doft delete and preserve the underlying attachment.
Note the downside of this is is that any delete callbacks are skipped.
At some point ActiveStorage might support `has_one_attached purge: false` or something, and at that point we can revert to using that.