Add min and max weights to dry_weights
Add minimum and maximum weights to clinical dry weights. It is down to the clinician to populate these. They might choose to take particular action (e.g. change drug) when the weight drops or rises below/above the min/max.
Validate that the difference between weight and the min or max is not > 90kg