Use toast notifications for flash notices
Display a temporary notification in the top right of the screen
when there is a flash message to display (e.g. 'Saved successfully'),
unless we are on the login page, where we display flash messages
e.g. 'Username or password not recognised' as inline html above
the username input.

Use a view component to encapsulate flash message logic.