Display count of unread messages next to username in top menu
If a user has unread messages, display the count of these next to the user's name in the top menu bar. If they have > 99 then display 99+.
Use a view component and a new conditional index to speed up fetching and rending the count.
No caching for now as that would require expiring the cache when a message is read or received, and would get a bit unneccessarily conplicated.