Introduce Topic dependent dynamic letter sections
via Topic#section_identifiers which would contain database changable
list of Section classes that conform to the Section interface.

The Section interface is represented using the old name Part in
`Letters::Part`

New additions to the interface are:

`identifier` - allows to identify the section not by a class name,
but by a symbol.

`position` - allows to staticly, or in the future dynamicly re-order
letter sections