Update pathology_observations_grouped_by_date SQL view to group by date note datetime
Previously this view would group pathology by observed_at (datetime) in order to separate out (in the UI on HD MDM) the pre and post urea results occuring on the same day so that both results were visible. However now we have the ability to rename the post URE as P_URE and generate the URR from pre and post URE automatically, we no longer need to have a row per datetime, which creates a lot of excess rows, and instead it makes more sense to have all results for a day on one row.