Display clinic name on clinic visit letters
If a letter is associated with a clinic visit (cv), we used to include only the cv date in the letter.
Having the clinic name and code on the letter (or just the name if code is blank or the same as the name) enables admin staff to make quicker decisions when processing the letter, eg if a manual workaround to
print or process the letter outside of Renalware.

Also display alongside name wherever we display a clinic eg AccessClinic / C123