Add a clinic filter on Renalware -> Letters
Filters the list to show only letters based on a clinic visit (the letter event) where the clinic visit clinic matches.

Amend the :long date format from 'Fri 14 Oct 2022' to 'Fri 14-Oct-2022' to be in line with NHS standards.

Add a configuration setting for the separator joining clinic name + code in lists and on letters.

Amend the clinic visit description line on the letter so that the clinic date never wraps
even if the clinic name and code strings are long.