Fix flaky system test in #record_prescription_for for good
> And records the prescription for the infection      # features/step_definitions/exit_site_infections_steps.rb:58
>      expected to find css "#new_medications_prescription" but there were no matches (RSpec::Expectations::ExpectationNotMetError)


Problem with the `medications/prescriptions.rb#record_prescription_for`
was not in the function itself, but in the one preceding it -
`pd/infection_organism#record_organism_for`


In `#record_organism_for` - the "Save" action was not properly "waited-up", and as a result,
and as a result, the `record_prescription_for` was clicking "successfully"
on a link, but the page was shortly after refreshed, when the previous "Save"
had the page refreshed.

Thereby, the call after

```
expect(page).to have_selector("#new_medications_prescription")
```

was the one failing.