Fix invalid authenticity token errors
Change `protect_from_forgery with: :exception` to `protect_from_forgery with: :reset_session` in ApplicationControllers.

The  invalid authenticity token error we were getting before was because when the user sits on the login page eg overnight and the session cookie expires, the next POST request to login results in Rails not beeing happy because the CRSF token in the HTML form is not symetric with the cookie value, so tampering is assumed. We were explictly raising this as an error (useful if you are interested in knowing you are being hacked) but as this occurs many times a day naturlally, its best to use the reset_session strategy which will now just for the login page to reload. The end result is that rather seeing a blank page or error page, the user after logging in after a long delay will have to re-enter their username and password (ie login in twice). Not ideal but better.

You can reproduce the error by loging out and then deleting the session cookie and trying to login again. Alternatively you can use the dev tools to change the value of the csrf token and then submit the form to get the same error.