Remove redundant DumbDelegator class
Now that we are using the DumbDelegator gem.
However I think we should remove the DumbDelegator gem and default to using delegate_missing_to as it is one less gem and the intention is a bit clearer. Its does mean we also have to delegate to_s manually however as this method is not really 'missing'!