Baseline letter Transfer Of Care work
This squashed commit represents the Transfer Of Care MESH work we did
before switching to GP Connect. We are still keeping the ToC work around
in case we switch back to that workflow later.
At this point we have a working ToC implementation with a few missing
bits of data in the letter, and had passed ITK assurance using this solution.

Some key commit comments of note:
- Add uuid to users so that we can reference them in globally and uniquely in ToC letters (etc) as "urn:uuid:#{user.uuid}"
- Add uuid to clinic_visits so that we can reference them in globally and uniquely in ToC letters (etc) as "urn:{uuid}"
- Support MESH mailbox lookup from ODS code
- Add thoughts on resolving mailbox from practice
- Resolve Practice MESH mailbox id if missing
- Mark NHS number as unverified as we are not using PDS
- Add uuid to hospital_centres
- Read TOC organisation ODS code and UUID from config