By default do not allow QR codes in letters
Renalware.config.allow_qr_codes_in_letters defaults to false.
The dummy app sets it to true so QR code insertion will work in development but not by default in any host app. This gives us a chance to asses the feature a bit more before releasing.
The admin UI for QR codes works regardless of the config setting.