✨  Create DM+D Drugs
Use data from a Terminology server at the NHS to populate the choices
of Drugs that a patient can be prescribed.

This following collections of data are controlled via DM+D:

* Drug name - Virtual Therapeutic Moiety (VTM)
* Trade Family
* Form
* Route

Supporting data includes:

* Virtual Medical Products (VMPs)
* Actual Medical Products (AMPs)

History
---

Previous, a list of Drugs was manually maintained. The DM+D implementation
replaces it.

Syncing
---

The DM+D data is synced on a regular basis with each hospital system.
There are no API calls occuring during the use of the system,
as all data is stored locally.

Codebase & entrypoint
---

Most of the code is located in the Drugs::DMD namespace. There are
explanations for the main classes in the codebase.
See:

    Renalware::Drugs::DMD::SynchroniserJob
    Renalware::Drugs::DMD::FullSynchroniser

See
---

* https://github.com/airslie/renalwarev2/wiki/DM-D-Drugs

* https://digital.nhs.uk/services/terminology-servers#quick-start-for-developers

* [dm+d code system in NHS Terminology Server - NHSBSA](https://www.nhsbsa.nhs.uk/sites/default/files/2022-04/dm%2Bd%20Code%20System%20Structure%20April%202022.docx)

* [Shrimp Browser](https://ontoserver.csiro.au/shrimp/?concept=138875005&valueset=http%3A%2F%2Fsnomed.info%2Fsct%3Ffhir_vs&fhir=https%3A%2F%2Fontology.nhs.uk%2Fproduction1%2Ffhir)

    -> Use `https://ontology.nhs.uk/production1/fhir` as server and login