Introduce a Hospitals::Department model
Currently we have
- Hospitals::Centre which represents a hospital
- Hospitals::Unit which represent a location where treatment (e.g. HD) is given.

This PR introduces a Hospitals::Department which represents a unit of adminstration within the hospital centre. It has a name (eg 'Renal Department') and an address, phone and email. A hospital center can have many departments.

Allow a department to be assigned to a Letters::Letterhead so that the relevant  contact details can be pulled from the department when rendering Transfer Of Care letter FHIR XML letter representations.