Add VND Risk Assessments
Add the ability to create and view Venal Needle Dislodgment (VND)
risk assessments.

The risk assessment comprises 4 questions (each having a score) which
combine to make an overall risk score.
See app/assets/pdf/renalware/VND-Risk-Assessment.pdf
for the specifications.

VND risk assessments are related to HD and Access, but more HD focussed
so they live in the HD namespace. They can however be created from both
the Access or HD dashboards.

We display the assessments
- as a list on the HD dashboard
- as a list on the Access dashboard
- the latest on the printed HD Session protocol
- In the HD MDM list (as a filter also there)
- ON the HD MDM page for a patient

When displayed as a list, toggling open the row shows the 4 composite
scores.