Add support for opentelemetry in the demo app
Currently experiemtnal and confined to the demo app, but added as a means to understand opentelemetry so we can enable in Azure Monitor hopefully soon. Having Sentry-style metrtics there will help us to find performance hotspots and improve the user experience, without the need to spin up and maintain an opensource Sentry docker container.