Fix error creating Medication Review event
Currently if you launch the new event form and choose an event type of (eg) Cinical Frailty (which loads in a custom field) then change the event type dropdown to Medication Review, then the Cinical Frailty custom field does not get replaced because there is an ajax error that the partial app/views/renalware/medications/reviews/_inputs.html.slim does not exist.

Here we create the missing (blank) partial to resolve that error.