Add a job to execute a SQL function and store the output
The job can for example be cron scheduled, and will execute the SQL fn
and log the 2 returned OUT params (records_added and records_updated)
(or an error) in an API Log event.

Add a rake task so this can be called from a traditional cron job,
as well as directly from the good_job cron scheduler.