Allow choosing a code group when viewing tabular pathology
When viewing historical, current or recent pathology, allow the current
default view (the 'default' Pathology::CodeGroup) to be changed to any
other code group (eg 'Transplant') that is not context_specific=true
(ie not one that is designed for use in letters for instance).

This works using a turboframe to refresh the pathology table.

Currently there is not much shared code here as the three pathology
views work slightly differently, but there is scope to refactor to
view components.

At some point we will extend to anywhere pathology is displayed,
including MDM pages.

Add a pathology_code_groups.title which will be the valoue displayed
in the dropdown lists as the current name value can contain eg 'hd_mdm'.

On the Current view, switch to using CSS grid for each pathology table
'subgroup', and use a stimulus controller to set the row span for
each table depending on the actual size of the table. This allows a sort
of masonry affect.

Add support for an optional accent colour for the left hand
border on current pathology subgroup to make a subset easier to
identify. The colour choice is an enum of tailwindcss colour names.

Add support for an optional title for current pathology subgroups to
make their purpose clearer, which appears at the top of a subgroup
table.

Move the HGB (currently orange), CRE, URE and POT (green) colour
configuration from hard-coded css classes into the
pathology_observation_descriptions.colour column.
The colour choice is an enum of tailwind colour names.

Whitelist CSS classes that can be generate via database configuration ie
the values of the enum_colour_name as the various forms into which they
can be interpolated eg border-l-red-300, bg-orange-100.