Fix column highlighting in pathology tables
A bug had emerged where, if the 'view' onto the pathology changed, eg from 'default' to 'HD', the column would not highlight when hovered-over (a feature to help the user identify the relevant column header in the table).
Also fix missing popup on blue triangle comments in the recent pathology view.