Add HD slot requests
These provide a means to capture a patient's need for a
dialyser slot, and to also capture when they were allocated a slot, or
when someone was removed from the list for example due to starting PD.
A patient can only be on the list on the list of active slot
requests once, but can appear multiple times on the historical tab.
The historical tab is only visible to admin+ users.