Use js to highlight 'general' patient menu items
Use a stimulus controller to add an active class to the selected menu item in the 'General' section in the LH patient menu/gutter. This allows us to cache the _general partial (invalidated when the patient is touched when eg a event is added etc, so counts update in the menu items) in order to save about 20ms per patient page impression.