Rollback previous attempt to handle an AJ race condition
Previously tried to do a search for the patient if it has been created by another thread in between our thread not-finding it and trying to create it. My solution was a bit misguided as in practice, at the point of rescuing the error, the transaction is already aborted - so we just end up with a different error. Will rely on a retry_on handler (#4523) instead as the race condition occurs fairly infrequently