Remove POS acronym
It beaks devise/responders in Rails 7 where the http :POST verb symbol is resolved as :POS_T. This then means that in the responders gem, when the Responder class is trying to resolve a template to render, after e.g. a validation error in the #create action, it looks up the template name in its DEFAULT_ACTIONS_FOR_VERBS const and where it should find DEFAULT_ACTIONS_FOR_VERBS[:post] => "new", it finds DEFAULT_ACTIONS_FOR_VERBS[:post_t] => nil and this results in the Responder class redirecting to GET /users which results in a 401 and then goes to the sign in page.