Changes to death modality warning
When the death modality is selected we currently display a message about prescriptions being terminated when they save. The trigger for this message is if the selected modality name is "Death". However BLT have renamed this to eg "DECEASED", so this change is to use the modality_description.code column (always "death") as the trigger, rather than the name.