Add n demo scheduled function to illustrate updating patient.send_to_renal_reg
A hospital may wish to run a scheduled eg nightly task to find any patients who should be sent to the Renal Reg but currently aren't, and set end_to_renalreg=true.
This demo function (which can be copied into a hospital's app and tweaked as required) finds all patients with send_to_renal_reg = false and where no explicit decision has been made to opt them out, and sets send_to_renal_reg = true if their modality is Transplant or HD or PD, or their modality is low_clearance or nephrology and they have an egfr < 30. It returns the number of rows updated so this info can be displayed in the API Logs table.