Break reporting into a package Step 1
Step 2 will unfort need to prefix a lot of  non-reporting urls with "renalware.", otherwise we do not break free of the reporting scope when clicking on menu items. We might perhaps just need to apply the prefix to menu items to start with, and any links to patients generated in reports.

As a reminder, this is a quick proof of concept to see if we can break the RW modules/namespaces into packages (starting with reports which we know has no dependencies) in order to discover/fix/prevent unhelpful cross domain calls which are making the app harder to work with.