Introduce experimental HL7 feed processing alongside current approach
This is a different approach with the advantage that we are not storing
multiple similar copies of an ORU message as updated arrive (ie
partial, partial, partial, final) but just store one with the most
recent MSH sent_at date.

Simply put:
- Mirth extracts 7 or 8 key bits of info from the HL7 message
  and calls a SQL function to do the heavy lifting
- the SQL function upserts into a messages table, always inserting if
  ADT, and upserting based on ORC Filler Order Number if an ORU (but
  only if its a newer message). The fn also adds an entry to a
  queue table, along with the message id, to flag that there is
  something to do. The idea here is to avoid polling the large messages
  table for new work.
- a cron-style scheduled active job polls the queue table every few
  minutes and creates individual ActiveJob (to process the individual
  HL7 message) for each row, then deletes that row.

The idea is to run this new mechanism alongside the current one in
order to find any issues and edge cases, before switching over to it.

The naming of models in this PR is a bit odd but I wanted to help avoid
accidental name clashes, and to make it clear it is experimental and
to be refactored at some point before used for real in production.

There are similarities here to a previously tried approach, but that had
a flaw in that messages relating to an ORU test could be processed out
of order by the various background workers. This solution solves that
issue.