Allow a permitted user to renew selected HD Prescriptions for eg 6 months
Add a new button to a patient's prescriptions view called 'Renew HD Prescriptions'.
Only a user with the HD prescriber role can see this.
When clicked it displays a dialog listing 'Give on HD' (non-stat) prescriptions that are
eligible for renewal. The user can de-select any that do not apply.
In clicking 'Renew', checked prescriptions are terminated and renewed
with the a prescribed_on of today and a terminated_on of today +
Renalware.config.auto_terminate_hd_prescriptions_after_period