Store OBX.11 Observation Result Status
Store the results status (eg P=Preliminary, F=Final) so that we can give precedence to Final results when displaying. Since a recent change to use OBR date for for the OBX values (rather than OBX observed_at), we need a better way of filtering out preliminary results which might say eg 'Sent to Lab'; a subsequent OBX with results_status of F=Final will arrive with the correct value, and we do not want the P result to obscure the F one.