Move to an explicit nested layout for letters/lists
As part of breaking out letters/lists_controller into several letters/lists/*_controller classes (to reduce complexity when adding new tabs to the Renal -> Letters page, eg for GP Connect letters), a useful first step is to remove use of the `within_non_patient_layout` helper, which is adding a bit too much magic and is hard to grok; instead the standard rails nested layout is a more well-documented approach.
Note the change to renalware.html layout to check for a :content block, which is the key to being able to create a sub layout.
https://guides.rubyonrails.org/layouts_and_rendering.html#using-nested-layouts