Remove unused keyboard shortcuts
Removed ctl+f (focus on Patient search) as it interferes with the browser search-on-page functionality
Removed ctl+m (send patient message) also - has not been advertised and I do not believe anyone is using it.

If we decide to re-introduce shortcuts we could use keyboard event actions introduced in stimulus 3.2.