Fix error creating a new HD Session that has PGDs assigned
Applies only when creating a new HD Session, adding one or more patient group directions (PGDs), and then clicking 'Create' or 'Save and Signoff'. Editing an existing session add adding PGDs at that point is not affected.

The issue was that `accepts_nested_attributes_for` was not saving the SessionPatientGroupDirection join records successfully because it could not work out how to assign the parent session_id to the record. It worked when doing an update, where the session_id is already known, but not during a create, where it has to assign the session_id to the SessionPatientGroupDirection after the session has been created.

The solution was to
- add inverse_of to the belongs_to :session in SessionPatientGroupDirection
- add a corresponding inverse_of on the has_many :session_patient_group_directions in Session
- change the validation in SessionPatientGroupDirection to validate presence of the session and pgd relations, not their ids.

This is a fairly well documented gotcha.