Fix HD Diary slot creation issue
If scope was the first one in the dropdown eg 'Dialysing Monday Eve' then after choosing a patient and clicking 'Add to this and all future weeks', the patient was only added to the Weekly diary not the Master one. This was because some javascript in the diary slot form had not  executed when the modal loaded.

The fix was to remove the javascript altogether as it was unnecessarily complicated, and to detect that the master diary is the target diary in the controller, by looking for the presence of params[:master] (the master btn in the form submit).