Supersede rather than directly update HD Profile on death modality
Previously on death modality the patient's current hd_profile was updated to null-out hospital_unit and schedule_definition. This was the wrong approach and we should instead deactivate the current profile and create a new one with those nulled values. This way makes it easier to find what unit a patient was at before they died.