Offer vaccines filtered by ATC code when adding Vaccination event
When adding a Vaccination event, we already show two custom fields:
- vaccination_type
- drug (vaccine)

Currently the list of drugs is always the same - the entire list of
drugs under the drug type 'vaccine'.

In this change we allow the drugs list to be pre-filtered to include
only those relevant to the selected vaccination type.

If the new column virology_vaccination_types.atc_codes (string[])
is populated then we join onto vmps by matching on any vmp.atc_code that
starts with any code in atc_codes[].