Allow for display of configured external links on Logon and dashboard pages
If the ENV var EXTERNAL_LINKS is in the format eg "title1^url1|title2^url2" then parse the title + url pairs and display each as external links on the login page, and as action buttons on the user's dashboard next to the dashboard title.

This is used at Barts to add an easily accessible link to the Qlick Sense Renalware Dashboard, but could be used to link to e.g. Power BI or other external applications.