Add a letter_snomed_document_types table
A letter topic can be assigned a snomed_document_type,which will be used in GP Connect FHIR messages as Bundle.documentType. If the letter topic has no assigned snomed_document_type, the snomed_document_type with a boolean default_type=true (of which there can be only one due to unique index) will be used.