Support page-specific help using 'tours'
Use shepherd.js to show a 'tour' for a specific page.

While the 'tours' approach to showing help to a user is generally
called 'onboarding', it is not the same as the onboarding process in
Renalware, which is the compulsory viewing of a series of introductory
videos before a user is allowed to use the system. See other PRs.

A tour in this context is a series of annotations displayed in a
specific order, each pointing to specific target element on a page,
in order to explain to a user what each page element does.
The text steps/annotations on each page are stored in the database.

To show a tour, you could:
- add a stimulus controller to the page, eg data-controller="tours"
  and it would only show the tour if the user has not already seen it
  (this is not yet implemented).
- add a Help link to the page, which when clicked would load and start
  the tour for the current page (this not is implemented).
- both approaches together.

Development on refining tours is ongoing, but paused while we focus
on the onboarding help process instead. So this PR is about
introducing the tours system, but not yet using it in the UI.