Display a list of the consultant's HD Prescriptions due to expire on the dashboard
If the user is a consultant and has been assigned as named_consultant to
to any HD patients who have auto-terminated prescriptions due to expire
within 14 days (configurable) or having expired within 14 days
(configurable) then show these in a component on their dashboard.
Allow the component to be hidden via configuration.
Component will only be visible if the hospital is configured to use the
auto terminating HD Prescriptions feature.