First phase in migrating marital status from enum to database
The intention is to migrate at some point from the old martial_status enum to a new db-backed marital_status in order for different marital statuses code sets to be supported across different hospitals.

This PR introduces the new table so we can start populating it with acceptable values at the different sites, and to asses whats involved in the migration.

In a future PR we will address changing patient.marital_status to point to the database-backed version