Change current_observation_set date handling
Ensure Europe/London timezone is applied to dates before inserting into current_observation_sets json hash. Without this, a result observed at say midnight on 21 Jun could appear in the UI as observed on the 20 Jun. This is because it will be stored in the database as UTC, and the time zone offset was not applied correctly when updating the current_observation_sets json.