Improve formatting of ORM messages when displaying in patient notes
An ORM HL7 message causes the patient to be added to Renalware. in the messages is an OBR with repeating OBX segments containing various key value pairs - these need to be concatenated into something sensible in the patient notes field. This PR tidies up the formatting by formatting dates and timestamps correcting, and formatting physician names sensibly